Refactor application folder structure
