--- shoppinglist_react/.travis.yml
@@ -1,8 +1,0 @@
-language: node_js
-node_js:
-  - "6.11.2"
-
-script:
-  - npm test
-
-after_success: 'npm run coveralls'
