+++ src/__tests__/App.test.js
@@ -5,7 +5,7 @@ import expect from 'expect';
 import App from "../components/App";
 import renderer from 'react-test-renderer';
 import {BrowserRouter} from 'react-router-dom';
-import "./setup_test";
+import "../setup_test";
 
 
 describe('App component', () => {
