+++ src/__tests__/email_send.test.js
@@ -1,8 +1,8 @@
 import moxios from 'moxios';
 import {mount, shallow} from 'enzyme';
 import React from 'react';
-import "./setup_test";
-import ResetPassword from '../components/auth/reset_password'
+import "../setup_test";
+import SendEmail from '../components/auth/email_send'
 
 
 describe('mocking email send requests', function () {
@@ -17,25 +17,25 @@ describe('mocking email send requests', function () {
             moxios.uninstall()
         });
 
-        it('stub response for reseting password', function (done) {
-            let url = JSON.parse('{"params":{"token":"sthii"}}');
-            const reset = mount(<ResetPassword match={url}/>);
+        it('stub response for successful email sent', function (done) {
+            const send_email = mount(<SendEmail/>);
 
-            reset.state().email = 'chris@gmail.com';
-            reset.state().new_password = '12345';
-            reset.state().confirm_password = '12345';
-
-            reset.instance().handleReset({
+            send_email.state().email = 'chris@gmail.com';
+            send_email.instance().contentHeader = () => {
+            };
+            send_email.instance().handleReset({
                 preventDefault() {
                 }
             });
-            moxios.stubRequest('http://127.0.0.1:5000/auth/reset_password/sthii', {
+
+            moxios.stubRequest('http://127.0.0.1:5000/auth/reset_password', {
                 status: 200,
-                response: {message: "Password has been reset"}
+                response: {message: "Email to reset password has been sent"}
+
             });
 
             moxios.wait(function () {
-                expect(reset.html()).toContain("Password has been reset");
+                expect(send_email.html()).toContain("Email to reset password has been sent");
                 done()
             })
 
@@ -43,8 +43,8 @@ describe('mocking email send requests', function () {
 
 
         it('Changes email state when on change event is called', () => {
-            const reset = shallow(<ResetPassword/>);
-            const inputEmail = reset.find("input[name='email']");
+            const send_email = shallow(<SendEmail/>);
+            const inputEmail = send_email.find("input[name='email']");
             const event = {
                 target: {
                     name: 'email',
@@ -55,7 +55,7 @@ describe('mocking email send requests', function () {
                 }
             };
             inputEmail.simulate('change', event);
-            expect(reset.state().email).toBe(event.target.value);
+            expect(send_email.state().email).toBe(event.target.value);
         });
     }
