+++ src/__tests__/login.test.js
@@ -2,7 +2,7 @@ import moxios from 'moxios';
 import {mount, configure,shallow} from 'enzyme';
 import React from 'react';
 import Adapter from 'enzyme-adapter-react-16';
-import "./setup_test";
+import "../setup_test";
 import Login from '../components/auth/login'
